Refactored to move some functions around to where they fit better
