+++ client/src/components/modal-forms/index.js
@@ -4,8 +4,9 @@ export * from './RejectedForm';
 export * from './HireForm';
 export * from './NewParticipantForm';
 export * from './EditParticipantForm';
-export * from './NewSiteForm';
+export * from './NewSiteDialog';
 export * from './EditSiteForm';
+export * from './NewPhaseDialog';
 export * from './PSIForm';
 export * from './EditPSIForm';
 export * from './CohortForm';
